blogs/blog-08-production-deep-dives/pillar-02-security-sso/code/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
